+++ packages/reakit/src/Avatar/__tests__/Avatar-test.tsx
@@ -1,0 +1,33 @@
+import * as React from "react";
+import { render } from "react-testing-library";
+import Avatar from "../Avatar";
+
+test("html attrs", () => {
+  const { getByAltText } = render(
+    <Avatar id="test" aria-label="test" alt="test" />
+  );
+  expect(getByAltText("test")).toHaveAttribute("id", "test");
+  expect(getByAltText("test")).toHaveAttribute("aria-label", "test");
+});
+
+test("styled", () => {
+  const { container } = render(<Avatar />);
+  expect(container.firstChild).toMatchInlineSnapshot(`
+.c0 {
+  margin: unset;
+  padding: unset;
+  border: unset;
+  background: unset;
+  font: unset;
+  font-family: inherit;
+  font-size: 100%;
+  box-sizing: border-box;
+  background-color: unset;
+  color: inherit;
+}
+
+<img
+  class="c0"
+/>
+`);
+});