+++ packages/website/src/utils/track.ts
@@ -1,4 +1,5 @@
 const track = (event: string, meta?: any) => () => {
+  // @ts-ignore
   dataLayer.push({ event, meta });
 };