+++ packages/reakit/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.15.8](https://github.com/reakit/reakit/compare/reakit@0.15.7...reakit@0.15.8) (2018-10-16)
+
+
+### Bug Fixes
+
+* Fix `will-change` css prop on `Hidden` ([50eff5c](https://github.com/reakit/reakit/commit/50eff5c))
+
+
+
+
+
+## [0.15.7](https://github.com/reakit/reakit/compare/reakit@0.15.6...reakit@0.15.7) (2018-10-09)
+
+
+### Bug Fixes
+
+* **typescript:** Remove `hide` from `TabsTab` prop types ([b551e41](https://github.com/reakit/reakit/commit/b551e41))
+
+
+
+
+
+## [0.15.6](https://github.com/reakit/reakit/compare/reakit@0.15.5...reakit@0.15.6) (2018-10-09)
+
+
+### Bug Fixes
+
+* **typescript:** Make `Group` props optionals ([f845933](https://github.com/reakit/reakit/commit/f845933))
+
+
+
+
+
 ## [0.15.5](https://github.com/reakit/reakit/compare/reakit@0.15.4...reakit@0.15.5) (2018-10-05)